backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
supabase==2.4.6
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0
